update request instruction, update Tmrt function
Previous iteration included use of a dummy directory for download so jobs would continue to be added to the queue without waiting for download. To alleviate confusion this approach has been shifted to a comment, and direct download is used.

A vectorized version of the Tmrt function is also used to avoid error that was coming up with raster-based application.